Add basic pages
Next JS has a pages subdirectory where you can just export a react
component and that's the whole page, I guess. Pretty cool. The name of
the file is the route. So the file `sell.js` is accessed at `/sell`,
also neat. I like small conventions like that. Also links are declared
with `<Link>` tags, not anchor tags.

At this point we're running the server in the `frontend` directory,
which surprised me. The code for this is run out of npm, you can see
what to run by looking in the `scripts` section of `package.json`.

`npm run dev` to run the server on port 7777