Add layout scaffolding
This was mainly adding layout files and components to the app. Layouts
in this context being kinda railsy, but basically just meaning things
that show up on every page. This was done by overwriting a folder next
js uses called _app.js, and wrapping the `Component` prop it's passed
with a `Page` component that I wrote. An interesting Meta component was
also written to encapsulate the logic used for meta tags, like encodings
and whatnot.