Add theme and layout styles with styled-components
Styled components are a new way people are adding styles to their react
components across the industry. The main advantage to it for me is that
instead of writing styles in a JS object that are camel-cased, you can
write literal CSS in your JavaScript, and even get a lot of the benefits
of a pre processor like Sass because these are powered by ES6 template
strings, which allow the use of interpolations, so variables and nesting
are built right in!

It's also worth mentioning that there's a huge variety of syntax
highlighters for different editors for this. I found a very good one for
Vim, Wes uses one for VS Code that he likes, and I'm sure others for
atom and the like are supported as well.

Other notes from the
[docs](https://www.styled-components.com/docs/basics#motivation):

* There's no more guessing with CSS names. Issues like wondering whether
or not a CSS name that you want to use is already being used somewhere,
or knowing whether or not it's safe to delete a class name are pretty
much fundamentally handled by this. When the JS gets compiled down to
CSS, the class names will be automatically generated.

* More control over where your CSS lives. It can live alongside your
JSX, or you can import it in a more traditional fashion.

* Has themeing support with a ThemeProvider component. This component is
designed to wrap around any components that you have already created
using styled-components, and provide those components with variables
(accessed through props) to use to interpolate the values for styling.