Add Global styles and Imported styles
The `injectGlobal` function from styled-components allows you to define
styles and have them injected everywhere in the app. This is a great
place to define any styles you want to have on the `html` and `body`
tags. The only downside is that you're inherently outside of the
`ThemeProvider` tag, so you can't interpolate themes from that without
importing it to wherever your `injectGlobal` function is being called
(unless you're in the same file anyways, which is the case right now).

There's a directory in this project called `/styles` that is housing all
the styles for various parts of the app so I don't have to write them.
These are files that literally import the styled-components package,
then define and export a `const` that can be used like a react element
in the JSX.