Add NProgress for site-wide progress bar
NProgress is a tool used to add a progress bar to SPA's that don't
reload the page, only change the contents of the same page and alter the
URL. Because we're not changing pages, browsers don't show any
indication that something is loading, so this is a remedy for that.

The tool hooks into the life cycle methods provided by Next's router,
and runs JS from there.