Install Prisma CLI
Installs the Prisma CLI and sets up a few configuration variables for
this. Right now the secret is commented out to allow for easier
development, but this wouldn't be acceptable in a non-educational
setting.

GraphQL is interesting. I'll probably edit this summary as I go but from
what I understand it's a *spec*. In this context, that means it's meant
to be the same across everything, in this case, website API's. The
problem it's trying to solve is that every RESTful API is totally
different, and you have to basically re-learn how to use them everytime
you want to get information from a third party like Github or Twitter.

Instead, now they're saying fuck it, it's one endpoint, for literally
every CRUD action. I'm not sure if this means you can do update and read
actions at the same time, but you definitely can from the same URL.
Also, GraphQL breaks down these actions in to _queries_ and _mutations_,
the former of which is _reading_ data, the latter of which is _changing_
it (so updating, creating, and deleting).

Prisma is an ORM, similar to ActiveRecord, except that it's written in
JavaScript and I guess works particularly well with GraphQL. Once nice
thing about it is that it's safely typed. I don't know TypeScript or
any strongly typed langauge, so I'm not 100% sure what all the
advantages are there, but I've heard good things about it.