Create basic queries and mutations
Queries and Mutations appear to be like Rails controllers. You can
define a query or mutation class in the `schema.graphql` file, and the
behavior of any functions you want that class to have.

For example, in this commit I've defined a function on the `Query`
resolver class called dogs. The syntax with which I've written this,
because graphQL is loosely typed, allows me to ensure that this function
can only return `Dog` objects. `Dog` objects are defined in this file
too, right now they're any object that has just a `name` attribute that
is a string, and the exclamation point ensures that it must be a string.
Similarly, the exclamation point on the `[Dog]!` array that defines the
return value of `dogs()` ensures that this array of `Dog` objects must
be returned.

Another thing you can do is write it as `[Dog!]!`. The added exclamation
point after the `Dog` class reference ensure that it must not be null.

If you define a function in the schema, such as `dogs` or `createDog`,
then you must create a corresponding function on the object's class or
the code will break.

These functions, as defined in the class, take 4 arguments:

`parent, args, ctx, info`

parent: allows for nested queries
args: whatever arguments were passed in to the query or mutation
      presented as an object
ctx: Short for context, this points to the context option that's set in
`createServer.js`, it's very important because this exposes the database
as well as any request information like headers or cookies.
info: More advanced cases, was not explained in detail in this step but
appears to be meta information about the query itself. Sounds like a
door to meta-programming sadness.

more fully explained here:
https://www.apollographql.com/docs/graphql-tools/resolvers/

I don't know where `global` came from and can't find out. :(